requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
itsdangerous
email-validator
pyotp
qrcode
python-dotenv
Werkzeug